fixed bug that removes the temp_dir on build error even if the cleanup: false

+++ mkdocs_multirepo_plugin/plugin.py
@@ -358,5 +358,5 @@ class MultirepoPlugin(BasePlugin):
             shutil.rmtree(str(self.temp_dir))
 
     def on_build_error(self, error):
-        if self.temp_dir:
+        if self.temp_dir and self.config.get("cleanup"):
             shutil.rmtree(str(self.temp_dir))
